05_hackathon/congestion_tracker/dashboard/requirements.txt
# requirements.txt generated by rsconnect-python on 2026-03-09 06:20:00.972263+00:00
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.128.0
fonttools==4.61.1
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
linkify-it-py==2.0.3
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.15.0
numpy==2.4.1
orjson==3.11.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
questionary==2.1.1
requests==2.32.5
seaborn==0.13.2
semver==3.0.4
setuptools==80.10.2
shiny==1.5.1
shinychat==0.2.8
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.4.0
websockets==16.0
